requirements.txt
numpy
streamlit
tensorflow-cpu
opencv-python
Pillow
pathlib
tflite_support
